bootable/userfastboot: Fix 64bit compilation error
Reviewed-on: https://android.intel.com/170716
Tested-by: Gumbel, Matthew K <matthew.k.gumbel@intel.com>

+++ microui/graphics.c
@@ -149,7 +149,7 @@ static int get_framebuffer(GGLSurface *fb)
     fb->width = vi.xres;
     fb->height = vi.yres;
     fb->stride = fi.line_length/PIXEL_SIZE;
-    fb->data = (void*) (((unsigned) bits) + vi.yres * fi.line_length);
+    fb->data = (void*) (((unsigned long) bits) + vi.yres * fi.line_length);
     fb->format = PIXEL_FORMAT;
     memset(fb->data, 0, vi.yres * fi.line_length);
